+++ src/main/kotlin/codegen/FlutterXPluginInfo.kt
@@ -3,10 +3,10 @@ package codegen
 object FlutterXPluginInfo {
     const val Version: String = "5.3.3.243"
     const val Changelog: String = """
-<h2>5.3.3 - 2025-03-08</h2>
+<h2>5.3.3 - 2025-03-10</h2>
 
-<ul><li>fix:</li></ul>
+<ul><li>fix: The problem of jitter in asset images exceeding a certain size (document preview)</li><li>fix: assets code generate tool setting errors</li><li>Some other problems solved</li></ul>
 """
 
-    const val BuildDate: String = "2025-03-08"
+    const val BuildDate: String = "2025-03-10"
 }