+++ src/main/kotlin/shop/itbug/fluttercheckversionx/form/socket/SocketRequestForm.kt
@@ -156,7 +156,6 @@ class SocketRequestForm(val project: Project, private val toolWindow: ToolWindow
         println("dispose.....SocketRequestForm")
         apiPanel.model.removeListDataListener(this)
         apiPanel.removeOnChangeConfigListen(this)
-        DioApiService.getInstance().dispose()
         super.dispose()
     }